raspberry-backend/requirements.txt
paho-mqtt==1.6.1
onvif-zeep==0.2.12
pymongo==4.6.1
python-dotenv==1.0.0 